bot/requirements.txt
pip
python-telegram-bot
textblob
python-dateutil
requests
